# service/apigateway/docs/swagger.yaml
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      retryable:
        type: boolean
      status:
        type: string
      trace_id:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
      validations:
        items:
          $ref: '#/definitions/errors.ValidationError'
        type: array
    type: object
  errors.ErrorType:
    enum:
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - CONFLICT
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - UNAVAILABLE
    type: string
    x-enum-varnames:
    - ErrorTypeInternal
    - ErrorTypeNotFound
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeTimeout
    - ErrorTypeUnavailable
  errors.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
...
  response.ApiResponseCategory:
    properties:
      data:
        $ref: '#/definitions/response.CategoryResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseCategoryAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseCategoryDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseCategoryDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.CategoryResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_category:
        type: string
      name:
        type: string
      slug_category:
        type: string
      updated_at:
        type: string
    type: object
  response.CategoryResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_category:
        type: string
      name:
        type: string
      slug_category:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/category-command/create:
    post:
      consumes:
      - multipart/form-data
      description: Create a new category with the provided details
      parameters:
      - description: Category name
        in: formData
        name: name
        required: true
        type: string
      - description: Category description
        in: formData
        name: description
        required: true
        type: string
      - description: Category slug
        in: formData
        name: slug_category
        type: string
      - description: Category image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created category
          schema:
            $ref: '#/definitions/response.ApiResponseCategory'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to create category
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new category
      tags:
      - Category Command
  /api/category-command/permanent/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a category record by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted category record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseCategoryDelete'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to delete category permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a category
      tags:
      - Category Command
  /api/category-command/permanent/all:
    delete:
      consumes:
      - application/json
      description: Permanently delete all trashed category records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all categories permanently
          schema:
            $ref: '#/definitions/response.ApiResponseCategoryAll'
        "500":
          description: Failed to delete categories permanently
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all trashed categories
      tags:
      - Category Command
  /api/category-command/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed category record by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored category
          schema:
            $ref: '#/definitions/response.ApiResponseCategoryDeleteAt'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to restore category
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed category
      tags:
      - Category Command
  /api/category-command/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all trashed category records
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all categories
          schema:
            $ref: '#/definitions/response.ApiResponseCategoryAll'
        "500":
          description: Failed to restore categories
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all trashed categories
      tags:
      - Category Command
  /api/category-command/trashed/{id}:
    post:
      consumes:
      - application/json
      description: Move a category record to trash by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully moved category to trash
          schema:
            $ref: '#/definitions/response.ApiResponseCategoryDeleteAt'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to move category to trash
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Move category to trash
      tags:
      - Category Command
  /api/category-command/update/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Update an existing category record with the provided details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category name
        in: formData
        name: name
        type: string
      - description: Category description
        in: formData
        name: description
        type: string
      - description: Category slug
        in: formData
        name: slug_category
        type: string
      - description: Category image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated category
          schema:
            $ref: '#/definitions/response.ApiResponseCategory'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Failed to update category
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing category
      tags:
      - Category Command
...
swagger: "2.0"